refactor(bridge): extract SwapRateDetails
